Add collapsible section to reply each comment
Add 2 new properties at CommentForm component, 'title' and 'action'
to show different titles and submit button text for each CommentForm.

Create a new component called CommentListCointainer to contain the main
CommentList component and this only contain the list of specifyied
comments.

Add a CommentForm and CommentList for each comment and its replies;
when is replied a comment, the form send the reply at form and then
send the parent comment and the reply to be added at the respectly
replies list path in database.