Replace ReplyForm by CommentForm
Remove the function that remove the spaces and convert to lowercase a
string, using as comment image a image link with the original username.

Using a CommentForm specify the type of CommentForm, in this case is a
reply, also send as parameter the title, tooltip title and button text.

Change the replies title state when are no one reply to a alert with
the text: 'Be the first at reply'.