Fix the view of multiline comments
The comment text was splited by each line break character '\n' and each
line was added in a div element with a br element inside.

Also it was changed the collapse element id from 'collapse' to
'replies'.

It was added a left border in CommentForm and CommentList elements and
a responsive size for this elements.